Rename <SettingsPanel/> to <VisualizationSettings/>
