+++ src/components/VisualizationSettings.svelte
